Started to implement the simulation

+++ ProcessorSimulation/IRegister.cs
@@ -6,9 +6,15 @@ using System.Threading.Tasks;
 
 namespace ProcessorSimulation
 {
+    /// <summary>
+    /// Immutable register in a processor.
+    /// Caution: The type of the value field is uint, but mostly byte values are stored.
+    /// This is because there exist special cases, with registers that need to store bigger values.
+    /// (E.g. MIP is a 12 bit register.)
+    /// </summary>
     public interface IRegister
     {
         Registers Type { get; }
-        byte Value { get; }
+        uint Value { get; }
     }
 }
